Refactor: separate functions from index.js
